requirements.txt
streamlit
PyPDF2
langchain-core
langchain-groq
cassio
python-dotenv
